+++ Sources/Scenes/SceneFactory.swift
@@ -42,16 +42,6 @@ class SceneFactory {
         return viewController
     }
 
-    static func createInviteFriendsListViewController(room: Room) -> InviteFriendsListViewController {
-        let viewController = InviteFriendsListViewController()
-        let presenter = InviteFriendsListPresenter(output: viewController)
-
-        let interactor = InviteFriendsListInteractor(output: presenter, api: APIClient(), room: room)
-        viewController.output = interactor
-
-        return viewController
-    }
-
     static func createAuthenticationViewController() -> AuthenticationViewController {
         let viewController = AuthenticationViewController()
         let presenter = AuthenticationPresenter(output: viewController)